feat: Organize helper modules under `util/`
This is intended to help clarify the folder structure in Hipcheck. We
have some helper logic that wraps `std` functionality with nicer error
messages, or extends iterators with helpful new methods, and those can
now be found under the `util/` directory.
