+++ hipcheck/src/analysis/metric/affiliation.rs
@@ -6,8 +6,8 @@ use crate::data::git::Commit;
 use crate::data::git::CommitContributorView;
 use crate::error::Error;
 use crate::error::Result;
-use crate::filesystem as file;
 use crate::hc_error;
+use crate::util::fs as file;
 use serde::de::Error as SerdeError;
 use serde::de::Visitor;
 use serde::Deserialize;
