+++ hipcheck/src/analysis/session/mod.rs
@@ -37,13 +37,13 @@ use crate::data::ModuleProviderStorage;
 use crate::data::PullRequestReviewProviderStorage;
 use crate::error::Error;
 use crate::error::Result;
-use crate::filesystem::create_dir_all;
 use crate::hc_error;
 use crate::report::Format;
 use crate::report::ReportParams;
 use crate::report::ReportParamsStorage;
 use crate::shell::Phase;
 use crate::shell::Shell;
+use crate::util::fs::create_dir_all;
 use crate::version::get_version;
 use crate::version::VersionQuery;
 use crate::version::VersionQueryStorage;
