+++ hipcheck/src/data/npm.rs
@@ -4,8 +4,8 @@ use crate::command_util::log_args;
 use crate::command_util::DependentProgram;
 use crate::context::Context;
 use crate::error::Result;
-use crate::filesystem as file;
 use crate::hc_error;
+use crate::util::fs as file;
 use pathbuf::pathbuf;
 use serde::Deserialize;
 use std::collections::HashMap;
