+++ hipcheck/src/util.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+
+//! Utility methods and types used throughout Hipcheck.
+
+pub mod fs;
+pub mod iter;
