+++ hipcheck/src/util/iter.rs
@@ -1,6 +1,25 @@
 // SPDX-License-Identifier: Apache-2.0
 
-//! A struct and trait for performing fallible filtering of Iterators.
+//! Iterator extension traits.
+
+/// A fallible analogue of the `Iterator::any` method
+pub trait TryAny: Iterator {
+	fn try_any<F, E>(&mut self, mut f: F) -> Result<bool, E>
+	where
+		F: FnMut(<Self as Iterator>::Item) -> Result<bool, E>,
+	{
+		for t in self {
+			match f(t) {
+				Ok(false) => continue,
+				result => return result,
+			}
+		}
+
+		Ok(false)
+	}
+}
+
+impl<I: Iterator> TryAny for I {}
 
 /// Represents an iterator and a fallible criterion for filtering it
 pub struct FallibleFilter<I, P, E>
@@ -62,7 +81,17 @@ impl<I: Iterator> TryFilter for I {}
 mod tests {
 	use super::*;
 
-	fn odd_not_three(n: &usize) -> Result<bool, String> {
+	fn odd_not_three(n: usize) -> Result<bool, String> {
+		if n == 3 {
+			Err(String::from("Error"))
+		} else if n % 2 == 1 {
+			Ok(true)
+		} else {
+			Ok(false)
+		}
+	}
+
+	fn odd_not_three_ref(n: &usize) -> Result<bool, String> {
 		if *n == 3 {
 			Err(String::from("Error"))
 		} else if *n % 2 == 1 {
@@ -73,12 +102,47 @@ mod tests {
 	}
 
 	#[test]
+	fn any_ok_true() {
+		let v = vec![2, 4, 5];
+
+		let result = v.into_iter().try_any(odd_not_three).unwrap();
+
+		assert!(result);
+	}
+
+	#[test]
+	fn any_ok_true_with_three() {
+		let v = vec![2, 4, 5, 3];
+
+		let result = v.into_iter().try_any(odd_not_three).unwrap();
+
+		assert!(result);
+	}
+
+	#[test]
+	fn any_ok_false() {
+		let v = vec![2, 4, 6, 8, 10];
+
+		let result = v.into_iter().try_any(odd_not_three).unwrap();
+
+		assert!(!result);
+	}
+
+	#[test]
+	#[should_panic]
+	fn any_err() {
+		let v = vec![2, 4, 3, 1, 5, 7, 9];
+
+		v.into_iter().try_any(odd_not_three).unwrap();
+	}
+
+	#[test]
 	fn filter_ok() {
 		let v = vec![1, 2, 4, 5];
 
 		let result = v
 			.into_iter()
-			.try_filter(odd_not_three)
+			.try_filter(odd_not_three_ref)
 			.collect::<Result<Vec<_>, String>>()
 			.unwrap();
 
@@ -91,7 +155,7 @@ mod tests {
 		let v = vec![2, 4, 6, 8, 3, 10];
 
 		v.into_iter()
-			.try_filter(odd_not_three)
+			.try_filter(odd_not_three_ref)
 			.collect::<Result<Vec<_>, String>>()
 			.unwrap();
 	}
